Normalize IOsonata USB source layout
